requirements.txt
asgiref==3.8.1
Django==4.2
django-cors-headers==4.7.0
djangorestframework==3.15.2
graphene==3.4.3
graphene-django==3.2.2
graphene-file-upload==1.3.0
graphql-core==3.2.6
graphql-relay==3.2.0
pillow==11.1.0
promise==2.3
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
six==1.17.0
sorl-thumbnail==12.11.0
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.12.2
